Add a way to run icu4x-datagen with cfg(test)
