+++ provider/datagen/src/lib.rs
@@ -133,6 +133,9 @@
 )]
 #![warn(missing_docs)]
 
+// See explanation in test_datagen.rs
+#![cfg_attr(icu4x_test_datagen, no_main)]
+
 mod driver;
 mod provider;
 mod registry;
